Add dual chain support for state router, prevent update when invoke
